Add methods to RentalSessionController

+++ src/main/java/com/senla/srs/controller/v1/RentalSessionController.java
@@ -5,6 +5,7 @@ import com.senla.srs.dto.RentalSessionRequestDTO;
 import com.senla.srs.dto.SeasonTicketDTO;
 import com.senla.srs.dto.UserResponseDTO;
 import com.senla.srs.mapper.RentalSessionRequestMapper;
+import com.senla.srs.mapper.RentalSessionResponseMapper;
 import com.senla.srs.model.RentalSession;
 import com.senla.srs.model.User;
 import com.senla.srs.service.RentalSessionService;
@@ -22,6 +23,7 @@ import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.NoSuchElementException;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Slf4j
@@ -33,6 +35,7 @@ public class RentalSessionController {
     private RentalSessionService rentalSessionService;
     private UserService userService;
     private RentalSessionRequestMapper rentalSessionRequestMapper;
+    private RentalSessionResponseMapper rentalSessionResponseMapper;
     private static final String NO_RENTAL_SESSION_WITH_ID = "No rental session with this serial number found";
 
     @GetMapping
@@ -41,13 +44,13 @@ public class RentalSessionController {
 
         if (userService.isAdmin(userSecurity)) {
             return rentalSessionService.retrieveAllRentalSessions().stream()
-                    .map(rentalSession -> rentalSessionRequestMapper.toDto(rentalSession))
+                    .map(rentalSession -> rentalSessionResponseMapper.toDto(rentalSession))
                     .collect(Collectors.toList());
         } else {
             try {
                 User authUser = userService.retrieveUserByAuthenticationPrincipal(userSecurity).get();
                 return rentalSessionService.retrieveAllRentalSessionsByUserId(authUser.getId()).stream()
-                        .map(rentalSession -> rentalSessionRequestMapper.toDto(rentalSession))
+                        .map(rentalSession -> rentalSessionResponseMapper.toDto(rentalSession))
                         .collect(Collectors.toList());
             } catch (NoSuchElementException e) {
                 log.error(e.getMessage(), NO_RENTAL_SESSION_WITH_ID);
@@ -88,7 +91,7 @@ public class RentalSessionController {
     private ResponseEntity<?> getById(Long id) {
         try {
             RentalSession rentalSession = rentalSessionService.retrieveRentalSessionById(id).get();
-            return ResponseEntity.ok(rentalSessionRequestMapper.toDto(rentalSession));
+            return ResponseEntity.ok(rentalSessionResponseMapper.toDto(rentalSession));
         } catch (NoSuchElementException e) {
             String errorMessage = NO_RENTAL_SESSION_WITH_ID;
             log.error(e.getMessage(), errorMessage);
@@ -98,12 +101,34 @@ public class RentalSessionController {
 
     @PostMapping
     @PreAuthorize("hasAuthority('rentalSessions:read')")
-    public ResponseEntity<?> createOrUpdate(@RequestBody RentalSessionRequestDTO rentalSessionRequestDTO) {
-        //ToDo Plug!!! Add access rights !!!
-        if (isValidCreateRentalSession(rentalSessionRequestDTO)) {
+    public ResponseEntity<?> createOrUpdate(@AuthenticationPrincipal org.springframework.security.core.userdetails.User userSecurity,
+                                            @RequestBody RentalSessionRequestDTO rentalSessionRequestDTO) {
+
+        Optional<RentalSession> optionalExistRentalSession =
+                rentalSessionService.retrieveRentalSessionByUserAndScooterAndBegin(rentalSessionRequestMapper
+                        .toEntity(rentalSessionRequestDTO));
+
+        if (optionalExistRentalSession.isEmpty()) {
             return create(rentalSessionRequestDTO);
         } else {
-            return new ResponseEntity<>("RentalSession is not valid", HttpStatus.FORBIDDEN);
+            RentalSession existRentalSession = optionalExistRentalSession.get();
+
+            if (existRentalSession.getEnd() != null) {
+                return new ResponseEntity<>("Rental session closed and not available for modification", HttpStatus.FORBIDDEN);
+            } else if (userService.isAdmin(userSecurity) || isThisUserRentalSession(existRentalSession, userSecurity)) {
+                return create(rentalSessionRequestDTO);
+            } else {
+                return new ResponseEntity<>("Changes to this Rental session are not available", HttpStatus.FORBIDDEN);
+            }
+        }
+    }
+
+    private ResponseEntity<?> create(RentalSessionRequestDTO rentalSessionRequestDTO) {
+        if (isValidCreateRentalSession(rentalSessionRequestDTO)) {
+            //ToDo Добавить расчет, различное поведение при налиичии даты окончания
+            return save(rentalSessionRequestDTO);
+        } else {
+            return new ResponseEntity<>("Rental session is not valid", HttpStatus.FORBIDDEN);
         }
     }
 
@@ -146,12 +171,12 @@ public class RentalSessionController {
     }
 
 
-    private ResponseEntity<?> create(RentalSessionRequestDTO rentalSessionRequestDTO) {
+    private ResponseEntity<?> save(RentalSessionRequestDTO rentalSessionRequestDTO) {
         RentalSession rentalSession = rentalSessionRequestMapper.toEntity(rentalSessionRequestDTO);
         rentalSessionService.save(rentalSession);
         try {
             RentalSession createdRentalSession = rentalSessionService.retrieveRentalSessionById(rentalSession.getId()).get();
-            return ResponseEntity.ok(rentalSessionRequestMapper.toDto(createdRentalSession));
+            return ResponseEntity.ok(rentalSessionResponseMapper.toDto(createdRentalSession));
         } catch (NoSuchElementException e) {
             String errorMessage = "The rental session is not created";
             log.error(e.getMessage(), errorMessage);
@@ -160,13 +185,16 @@ public class RentalSessionController {
     }
 
 
-    //ToDo Admin can only delete non-closed sessions
     @DeleteMapping("/{id}")
     @PreAuthorize("hasAuthority('rentalSessions:write')")
     public ResponseEntity<?> delete(@PathVariable Long id) {
         try {
-            rentalSessionService.deleteById(id);
-            return new ResponseEntity<>("Rental session with this id was deleted", HttpStatus.ACCEPTED);
+            if (rentalSessionService.retrieveRentalSessionById(id).get().getEnd() != null) {
+                rentalSessionService.deleteById(id);
+                return new ResponseEntity<>("Rental session with this id was deleted", HttpStatus.ACCEPTED);
+            } else {
+                return new ResponseEntity<>("Rental session closed and cannot be deleted", HttpStatus.FORBIDDEN);
+            }
         } catch (NoSuchElementException e) {
             String errorMessage = "A rental session with this id was not detected";
             log.error(e.getMessage(), errorMessage);
